Add support for sanitizers, matchers
Resolves #1877. This also makes it possible to compile `azure_core_test` for the wasm32 architecture but will not currently work for wasm32. I opened #2009 to track.
