+++ sdk/core/azure_core_test/src/recording.rs
@@ -3,27 +3,93 @@
 
 //! The [`Recording`] and other types used in recorded tests.
 
-#[cfg(not(target_arch = "wasm32"))]
-use crate::proxy::Proxy;
-use azure_core::test::TestMode;
+use crate::{
+    proxy::{policy::RecordingPolicy, Proxy},
+    Matcher, Sanitizer,
+};
+use azure_core::{test::TestMode, ClientOptions};
 use std::sync::Arc;
 use tracing::Span;
 
 /// Represents a playback or recording session using the [`Proxy`].
+#[derive(Debug)]
 pub struct Recording {
-    #[allow(dead_code)]
-    #[cfg(not(target_arch = "wasm32"))]
-    pub(crate) proxy: Arc<Proxy>,
-    #[allow(dead_code)]
-    pub(crate) span: Span,
-    pub(crate) mode: TestMode,
+    test_mode: TestMode,
+    // Keep the span open for our lifetime.
+    _span: Span,
+    _proxy: Option<Arc<Proxy>>,
 }
 
 impl Recording {
+    pub(crate) fn new(test_mode: TestMode, span: Span, proxy: Option<Arc<Proxy>>) -> Self {
+        Self {
+            test_mode,
+            _span: span,
+            _proxy: proxy,
+        }
+    }
+
+    /// Adds a [`Sanitizer`] to sanitize PII for the current test.
+    pub fn add_sanitizer<S: Sanitizer>(&self, _sanitizer: S) -> azure_core::Result<()> {
+        todo!()
+    }
+
+    /// Instruments the [`ClientOptions`] to support recording and playing back of session records.
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use azure_core_test::{recorded, TestContext};
+    ///
+    /// # struct MyClient;
+    /// # #[derive(Default)]
+    /// # struct MyClientOptions { client_options: azure_core::ClientOptions };
+    /// # impl MyClient {
+    /// #   fn new(endpoint: impl AsRef<str>, options: Option<MyClientOptions>) -> Self { todo!() }
+    /// #   async fn invoke(&self) -> azure_core::Result<()> { todo!() }
+    /// # }
+    /// #[recorded::test]
+    /// async fn test_invoke(ctx: TestContext) -> azure_core::Result<()> {
+    ///     let recording = ctx.recording();
+    ///
+    ///     let mut options = MyClientOptions::default();
+    ///     ctx.instrument(&mut options.client_options);
+    ///
+    ///     let client = MyClient::new("https://azure.net", Some(options));
+    ///     client.invoke().await
+    /// }
+    /// ```
+    pub fn instrument(&self, options: &mut ClientOptions) {
+        options.per_try_policies.push(Arc::new(RecordingPolicy {
+            test_mode: self.test_mode,
+        }));
+    }
+
+    /// Sets a [`Matcher`] to compare requests and/or responses.
+    pub fn set_matcher(&self, _matcher: Matcher) -> azure_core::Result<()> {
+        todo!()
+    }
+
+    /// Skip recording the request body, or the entire request and response until the [`SkipGuard`] is dropped.
     pub fn skip(&self, _skip: Skip) -> SkipGuard<'_> {
-        // Tell transport to start sending `x-recording-skip` header with `skip.value()`.
+        // TODO: Tell transport to start sending `x-recording-skip` header with `skip.value()`.
         let _header_value = _skip.value();
         SkipGuard(self)
+    }
+
+    /// Gets the current [`TestMode`].
+    pub fn test_mode(&self) -> TestMode {
+        self.test_mode
+    }
+
+    /// Gets the named variable from the environment or recording.
+    pub fn var(&self, name: impl AsRef<str>) -> Option<String> {
+        if self.test_mode == TestMode::Live {
+            return std::env::var(name.as_ref()).ok();
+        }
+
+        // TODO: attempt to get it from the recording or fallthrough to the environment; or, do we need separate calls like .NET to fallthrough?
+        todo!()
     }
 }
 
@@ -48,11 +114,13 @@ impl Skip {
 }
 
 /// When the `SkipGuard` is dropped, recording requests and responses will begin again.
+///
+/// Returned from [`Recording::skip()`].
 pub struct SkipGuard<'a>(&'a Recording);
 
 impl Drop for SkipGuard<'_> {
     fn drop(&mut self) {
-        if self.0.mode == TestMode::Record {
+        if self.0.test_mode == TestMode::Record {
             // TODO: Tell transport to stop sending `x-recording-skip` header.
             todo!()
         }
